Refactor page
