+++ src/app/page.tsx
@@ -1,47 +1,15 @@
 "use client";
-import bs58 from "bs58";
-import { Noto_Sans_JP } from "next/font/google";
+
 import { useSearchParams } from "next/navigation";
-import { FC, ReactNode, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import Footer from "@/components/footer";
 import Header from "@/components/header";
 import { CompleteModal, DownloadModal } from "@/components/modals";
+import Panel from "@/components/panel";
 import { SelectionView, TraineeView } from "@/components/views";
+import { decodeSelection, encodeSelection } from "@/utils";
 
-const noto_sans_jp = Noto_Sans_JP({
-  subsets: ["latin", "latin-ext"],
-  display: "swap",
-});
-
-type PanelProps = {
-  children: ReactNode[]|ReactNode|string;
-}
-
-const Panel: FC<PanelProps> = ({ children }) => {
-  return (
-    <div
-      className="w-full sm:w-[480px] border border-gray-100 shadow
-      rounded-2xl text-sm sm:text-base overflow-hidden"
-    >{children}</div>
-  )
-};
-
-const encodeSelection = (selected: number[]): string => {
-  return bs58.encode(selected);
-};
-
-const decodeSelection = (code: any): number[]|undefined => {
-  if (typeof code === "string") {
-    try {
-      return Array.from(bs58.decode(code));
-    } catch {
-      return undefined;
-    }
-  }
-  return undefined;
-};
-
-const EMPTY = [-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1];
+const EMPTY = [255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255];
 
 export default function Home() {
   const searchParams = useSearchParams();
@@ -65,15 +33,15 @@ export default function Home() {
   }, [selectionCode]);
 
   return (
-    <main className={`${noto_sans_jp.className} h-full`}>
+    <main className="h-full">
       <Header />
       <div className="my-12 sm:my-20 px-4 text-center">
         <h2
-          className="text-pd-pink-400 text-sm sm:text-xl font-bold mb-6"
+          className="text-pd-pink-400 text-base sm:text-xl font-bold mb-6"
         >PRODUCE 101 JAPAN THE GIRLS RANKER
         </h2>
         <div
-          className="text-pd-gray-400 text-center text-xs sm:text-base"
+          className="text-pd-gray-400 text-center text-sm sm:text-base"
         >
           初めて羽ばたいた101人の少女たちが
           <br className="sm:hidden" />
